+++ src/simulator.h
@@ -7,7 +7,6 @@
 using namespace std;
 
 class Simulator : public Parser {
-
 public:
   Simulator() {
     wires["1\'b0"] = new Based("1\'b0", NET);
@@ -16,7 +15,7 @@ public:
   ~Simulator(){};
 
   void generate_output(map<string, vector<bitvec>> &);
-  void generate_input(int num = 1000, string fixed = "", int val = -1);
+  void generate_input(int num = 1000, string fixed = "", long long val = -1);
   void read(const string &);
   pair<string, int> find_smallest_input();
   vector<pair<string, int>> input_info;
